Initial commit; labs 1 and some of 2
The state of the lab 1 code is stable here. It compiles, (probably) runs, and
is stable. The lab 2 code is exactly what I had written down after the end of
the lab section today, and needs some refinements with the velocity and whatnot
to bring it into line. A good refinement would also be to make it measure
rotation in a smarter fashion involving rotational velocity.